Add client to pubsubPublishMessage where possible
This strange change exists so Redis tcl tests work.

The Redis tcl test suite runs clients on DB 9.  We have
to inform the pubsub scripts about which DB the client was using
when PUBLISH was called.  The script client gets updated with
the DB of the PUBLISH caller (or back to DB 0 if no publish
client existed) so when the script runs later, it runs
under the context of the correct DB.

The only way to update the pubsub
script with each PUBLISH call is to put the caller/publisher
into the PublishMessage command.

Lua scripts can also SELECT their DB internally, but that
requires, at least for testing, leaking implementation
details of the test into the script.  The DB selected in an
async script is not persisted on the next call.  Every call
has the DB reset to the DB of the caller (or DB 0 if the
caller is internal to Redis (NULL)).

It's not pretty, but it works.