Add automated testing for creating clusters
This script creates many different cluster configurations
by varying the number of instances and replica counts
defined at cluster initialization time.  If any cluster
topology fails to converge successfully, the entire test fails.

This script currently works on my OS X system and in a Linux VM.
It's designed to work on a CI server so it just runs
in the background and the system will report if the script
fails.