Cluster: Include myself IP in nodes when available
Previously, the node you ran CLUSTER NODES on always reported itself
as IP:Port = ":0" in the node output.  The client requesting
the current node map would have to parse "myself" from the
flags and replcae the IP:Port with the IP:Port of the instance used
to retrieve CLUSTER NODES.  It seemed a bit convoluted.

If the user sets a bind address in redis.conf, we can use that for
the IP of this instance.  If the user doesn't set a bind address,
we fall back to the original ":0" behavior.