Allow FQDN as SLAVEOF without breaking Sentinel
Sentniel verifies configurations by IP, but if a hostname
gets set as master_host, Sentniel re-syncs to the already
connected master because it can't match the hostname to the IP address.

The downside to this approach is the DNS resolve is blocking
(we don't have an async DNS resolver), so it can potentially
stall the server while the lookup happens.

Fixes #2075
Fixes #2032